+++ packages/gatsby/src/services/write-out-redirects.ts
@@ -1,10 +1,10 @@
-import { IBuildContext } from "./"
 import reporter from "gatsby-cli/lib/reporter"
 import { writeRedirects } from "../bootstrap/redirects-writer"
+import { IDataLayerContext } from "../state-machines/data-layer/types"
 
 export async function writeOutRedirects({
   parentSpan,
-}: Partial<IBuildContext>): Promise<void> {
+}: Partial<IDataLayerContext>): Promise<void> {
   // Write out redirects.
   const activity = reporter.activityTimer(`write out redirect data`, {
     parentSpan,